requirements.txt
Django==4.2
httpx
python-dotenv
